+++ Runtime/Pages/SceneNameDependentScriptableObject.cs
@@ -1,0 +1,11 @@
+﻿using UnityEngine;
+
+namespace UAppToolKit.Core.Pages
+{
+    public abstract class SceneNameDependentScriptableObject : ScriptableObject
+    {
+#if UNITY_EDITOR
+        public abstract void AnySceneNameChanged();
+#endif
+    }
+}